+++ hahomematic/async_support.py
@@ -6,15 +6,17 @@ import asyncio
 from collections.abc import Callable, Collection, Coroutine
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures._base import CancelledError
+from functools import wraps
 import logging
 from time import monotonic
-from typing import Any, Final, TypeVar
+from typing import Any, Final, ParamSpec, TypeVar
 
 from hahomematic.const import BLOCK_LOG_TIMEOUT
 from hahomematic.exceptions import HaHomematicException
-from hahomematic.support import reduce_args
+from hahomematic.support import debug_enabled, reduce_args
 
 _LOGGER: Final = logging.getLogger(__name__)
+_P = ParamSpec("_P")
 _R = TypeVar("_R")
 _T = TypeVar("_T")
 
@@ -121,3 +123,31 @@ class Looper:
 def cancelling(task: asyncio.Future[Any]) -> bool:
     """Return True if task is cancelling."""
     return bool((cancelling_ := getattr(task, "cancelling", None)) and cancelling_())
+
+
+def loop_check(func: Callable[_P, _R]) -> Callable[_P, _R]:
+    """Annotation to mark method that must be run within the event loop."""
+
+    _with_loop: set = set()
+
+    @wraps(func)
+    def wrapper_loop_check(*args: _P.args, **kwargs: _P.kwargs) -> _R:
+        """Wrap loop check."""
+        return_value = func(*args, **kwargs)
+
+        try:
+            asyncio.get_running_loop()
+            loop_running = True
+        except Exception:
+            loop_running = False
+
+        if not loop_running and func not in _with_loop:
+            _with_loop.add(func)
+            _LOGGER.warning(
+                "Method %s must run in the event_loop. No loop detected.", func.__name__
+            )
+
+        return return_value
+
+    setattr(func, "_loop_check", True)
+    return wrapper_loop_check if debug_enabled() else func